Create work history endpoints.

+++ src/employees/services/work_history_service.py
@@ -1,31 +1,31 @@
 from typing import List
-from sqlmodel import select
+from fastapi import HTTPException, status
 from src.database.core import DatabaseSession
 from src.schemas.entities import WorkHistory
 from src.schemas.work_history_models import WorkHistoryRequest
+import src.employees.services.utils as utils
 
 
-def get_single_work_history_by_id(
-    db: DatabaseSession, employee_id: int, work_history_id: int
-) -> WorkHistory:
-    return db.exec(
-        select(WorkHistory)
-        .where(WorkHistory.id == work_history_id)
-        .where(WorkHistory.employee_id == employee_id)
-    ).first()
+def get_work_history(db: DatabaseSession, employee_id: int) -> List[WorkHistory]:
+    employee = utils.get_employee_by_id(db, employee_id)
+    work_history = utils.get_work_history_of_employee(db, employee_id)
 
-
-def get_work_history_of_employee(
-    employee_id: int, db: DatabaseSession
-) -> List[WorkHistory]:
-    return db.exec(
-        select(WorkHistory).where(WorkHistory.employee_id == employee_id)
-    ).all()
+    if employee is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Employee not found."
+        )
+    return work_history
 
 
 def create_work_history(
     db: DatabaseSession, employee_id: int, work_history: WorkHistoryRequest
 ) -> WorkHistory:
+    employee = utils.get_employee_by_id(db, employee_id)
+    if employee is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Employee_id not found.",
+        )
     history = WorkHistory(
         employee_id=employee_id,
         job_title=work_history.job_title,
@@ -43,11 +43,17 @@ def create_work_history(
 def update_work_history_register(
     db: DatabaseSession,
     employee_id: int,
-    work_history_employee: int,
+    work_history_id: int,
     work_history_changes: WorkHistoryRequest,
 ) -> WorkHistory:
-    register = get_single_work_history_by_id(db, employee_id, work_history_employee)
+    employee = utils.get_employee_by_id(db, employee_id)
+    register = utils.get_single_work_history_by_id(db, employee_id, work_history_id)
 
+    if register is None or employee is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Work history_id or employee_id not found.",
+        )
     for attr, value in work_history_changes.model_dump(exclude_unset=True).items():
         if hasattr(register, attr):
             setattr(register, attr, value)
@@ -61,6 +67,13 @@ def update_work_history_register(
 def delete_work_history_register(
     db: DatabaseSession, employee_id: int, work_history_id: int
 ) -> None:
-    register = get_single_work_history_by_id(db, employee_id, work_history_id)
+    employee = utils.get_employee_by_id(db, employee_id)
+    register = utils.get_single_work_history_by_id(db, employee_id, work_history_id)
+
+    if register is None or employee is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Work history_id or employee_id not found.",
+        )
     db.delete(register)
     db.commit()
